requirements.txt
asyncio
colorama
clickatell
pandas
pymongo
python-binance
numpy
TA-Lib
